Fix typo in workflow-example
