+++ server/example/workflow-example/src/main/java/com/eclipsesource/glsp/example/workflow/WorkflowDiagramManager.java
@@ -45,13 +45,13 @@ import org.eclipse.sprotty.SGraph;
 import org.eclipse.sprotty.SLabel;
 import org.eclipse.sprotty.SModelElement;
 
-import com.eclipsesource.glps.server.diagram.AbstractDiagramManager;
 import com.eclipsesource.glsp.api.types.EdgeTypeHint;
 import com.eclipsesource.glsp.api.types.NodeTypeHint;
 import com.eclipsesource.glsp.example.workflow.schema.ActivityNode;
 import com.eclipsesource.glsp.example.workflow.schema.Icon;
 import com.eclipsesource.glsp.example.workflow.schema.TaskNode;
 import com.eclipsesource.glsp.example.workflow.schema.WeightedEdge;
+import com.eclipsesource.glsp.server.diagram.AbstractDiagramManager;
 
 public class WorkflowDiagramManager extends AbstractDiagramManager {
